Jeffeb3 Layout
I'm really trying to keep a separation between information and instructions.

For example, Endstops are in the instructions, because they need to be set up (on machines that have them) and they are in instructions, to help people understand when they are useful, what kinds there are, and what the different benefits are. I am imagining people going through the instructions without making many decisions, and when they need to make a choice, we will point them to the information in the other part of the documentation.